+++ src/common/js/configure.js
@@ -7,7 +7,6 @@ import _ from 'lodash';
 import { ReactiveVar } from 'meteor/reactive-var';
 
 let _conf = {};
-
 TenantsManager._conf = new ReactiveVar( _conf );
 
 TenantsManager._defaults = {
@@ -36,12 +35,23 @@ TenantsManager._defaults = {
  */
 TenantsManager.configure = function( o ){
     if( o && _.isObject( o )){
-        _conf = _.merge( TenantsManager._defaults, _conf, o );
-        TenantsManager._conf.set( _conf );
-        // be verbose if asked for
-        if( _conf.verbosity & TenantsManager.C.Verbose.CONFIGURE ){
-            //console.log( 'pwix:tenants-manager configure() with', o, 'building', TenantsList._conf );
-            console.log( 'pwix:tenants-manager configure() with', o );
+        // check that keys exist
+        let built_conf = {};
+        Object.keys( o ).forEach(( it ) => {
+            if( Object.keys( TenantsManager._defaults ).includes( it )){
+                built_conf[it] = o[it];
+            } else {
+                console.warn( 'pwix:tenants-manager configure() ignore unmanaged key \''+it+'\'' );
+            }
+        });
+        if( Object.keys( built_conf ).length ){
+            _conf = _.merge( TenantsManager._defaults, _conf, built_conf );
+            TenantsManager._conf.set( _conf );
+            // be verbose if asked for
+            if( _conf.verbosity & TenantsManager.C.Verbose.CONFIGURE ){
+                //console.log( 'pwix:tenants-manager configure() with', o, 'building', TenantsList._conf );
+                console.log( 'pwix:tenants-manager configure() with', built_conf );
+            }
         }
     }
     // also acts as a getter